Update files
This looks awesome. I'm suggesting a few changes - in particular, I think the tibble exercise may be a bit confusing for new R users because it uses strange naming conventions, so I've swapped that with a more conventional example. Let me know what you think. I've also added a line to allow it to be graded.